First pass at isolating dependencies
