+++ .github/workflows/linting.yml
@@ -30,7 +30,7 @@ jobs:
         sudo apt-get install librdkafka-dev # For antares-client
         python -m pip install --upgrade pip
         pip install .
-        pip install .[dev,data-generation,samplers,plotting]
+        pip install .[dev,data-generation,sampling,plotting]
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Analyze code with linter
       run: |
